chore: move `hooks` to `lib` directory
